cpu/nrf5x: implement wdt peripheral driver

+++ tests/periph_wdt/main.c
@@ -32,8 +32,8 @@ int get_range(int argc, char **argv)
 {
     (void)argc;
     (void)argv;
-    printf("lower_bound: %d upper_bound: %d \n", NWDT_TIME_LOWER_LIMIT,
-           NWDT_TIME_UPPER_LIMIT);
+    printf("lower_bound: %d upper_bound: %"PRIu32" \n", NWDT_TIME_LOWER_LIMIT,
+           (uint32_t)NWDT_TIME_UPPER_LIMIT);
     return 0;
 }
